pyproject.toml
[project]
name = "voice-to-dynamics"
version = "0.1.0"
description = "Voice assistant for Microsoft Dataverse / Dynamics 365 (Swedish)"
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech>=1.44.0",
    "anthropic>=0.40.0",
    "fastmcp>=2.0.0",
    "msal>=1.31.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "sounddevice>=0.5.0",
    "numpy>=2.0.0",
    "websockets>=13.0",
    "openai>=1.54.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src", "mcp_servers"]
